add the client libraries to the repository
little note, the hbs plugin for requirejs has had to little bugs. the
first is, it is strange, that the plugin has to load new versions of
existing loaded libraries of underscore, handlebars. the second is, it
is not necessary to load the hbs file with mimetype xml, text/plain is
enough. this change prevents the xml parser error.