ConvexBottom bugfix
the points which are used to build the bottom convex hull should lie
under a line between the left and the right most points. to test only
if the point lies under the horizontal line with y value of the higher
y value of the two right or left most points was not enough.