BouncingVertices implement reflex filter
this filter does not work properly. with one reflex point it should
work. more then one can cause in 9 of 10 cases a endless loop. the
problem lies in the fact, that it could be possible to create a
situation where the current point could not be placed without
violating the simple polygon requirement and fulfil the reflex filter.