BouncingVertices
changed the bouncing radius behaviour. if the value is set to 0 the
value is calculated from the half of the segment length. this is used
to have short bouncing moves. which should improve the successfull
termination behaviour.

the calculation for the segment length is done in RegularPolygon and
then it is bypassed throw FixLocalOrientation to the BouncingVertices
algorithm.