FixLocalOrientation bugfix
the reflex nodes were calculated by rotated the vector l from point L
to point P. the rotation angle was choosen randomly from range 0.1 to
iota. iota is the remaining angle where new reflex points could be
placed. the problem was, that iota was decreased by the choosen rotation
angle, and then iota could be go below zero. this leads to a rotation
angle that produce points that were not more in correct order and leads
to segments that cross other segments.

the rotation angle is now calculated as fraction between iota and the
choosen reflex points + 1 for this segment.