add parameter --reflex-chain-max
with this value it is possible to regulate how long the reflex chains
should be. but it does not protect the possibility that one chain
could be really long. the problem is, if the fraction between nodes /
reflex nodes is small then there is not enough space to have small
reflex chains.