FixLocalOrientation bugfix
the distribution of the reflex nodes was not implemented correctly. it
is easier to implement the distribution with the new init function.
first the reflex points where distributed over the segments, and then
the reflex points were calculated in a second loop.

this change may also simplify further improvements for different
reflex node distributions.