FixLocalOrientation add stretch feature
the convex points were stretched outwards. a convex chain has to have
more then one elements, to be stretched. the convex points near the
reflex points were not stretched. the move outwards is of the length
of the segment length. for the moment it is only a scale of 1
implemented.