SpacePartitioning bugfix
there must be more then one bug, but it is difficult to name it.

one point was that the line direction in calculateRandomLine should be
changed if s_f lies on the negative side not on positive side.

it is also better to not include the points from the line to the S_f
and S_l in the for loop by excluding it with p != s respectively p !=
s_f and p != s_l

a improvement is to use the final_list directly to build the polygon
and not go over two independent lists.