BouncingVertices
add second option to calculate the bouncing point. the first was to calculate a
random point and check if the point is valid. the second option now is to
calculate a random line and the valid interval to move the point around on this
line, and then calculate a random point on this line.

the implementation is not finished. the main idea is implemented, but there are
some bugs to fix. one big is known, but some are still unknown.