refactored
replaced CommonSettings and Filters by the docopt std::map. the Filter and
CommonSetting approach was introduced because of the idea to use it on a gui. to
define the filters and the settings in cpp files and publish them to the browser
gui. but this part was not developed further. therefore it is removed. the gui
has to be rewriten completly. maybe with a new concept (e.g. pwa,
web-components, mvc)